Show soft login after screen turn off and on if required
